+++ tests/package-common/src/schema/system-tables.factory.test.ts
@@ -40,3 +40,33 @@ describe('state system table factory', () => {
     expect(State.SQLite.getTableBackendId(systemTablesB.schemaMetaTable)).toBe('b')
   })
 })
+
+describe('table backend id fail-fast', () => {
+  test('getTableBackendId throws for unassigned table defs', () => {
+    const tableDef = State.SQLite.table({
+      name: 'unregistered',
+      columns: {
+        id: State.SQLite.text({ primaryKey: true }),
+      },
+    })
+
+    expect(() => State.SQLite.getTableBackendId(tableDef)).toThrow(/not assigned to a backend/i)
+  })
+
+  test('getTableBackendId returns backend id after makeBackend tags it', () => {
+    const tableDef = State.SQLite.table({
+      name: 'registered',
+      columns: {
+        id: State.SQLite.text({ primaryKey: true }),
+      },
+    })
+
+    State.SQLite.makeBackend({
+      id: 'a',
+      tables: { tableDef },
+      materializers: {},
+    })
+
+    expect(State.SQLite.getTableBackendId(tableDef)).toBe('a')
+  })
+})
